Removing :en in favor of default locale
Hi there,

i have an app without english as available locale. So i got an error when we try to inspect something like 1.day. This is done automatically when we use the dalli cache.

I would like to change the :en to ::I18n.default_locale to be sure that this is always constant and is an available locale.

Tests are all green with this change.

Calculating -------------------------------------
:locale => :en                              2.024k i/100ms
:locale => ::I18n.default_locale   2.236k i/100ms
-------------------------------------------------
:locale => :en                                  25.758k (±26.3%) i/s -    117.392k
:locale => ::I18n.default_locale       26.311k (±18.1%) i/s -    127.452k 


Is there any other intention why this should be :en?